Added functionality to Add, Clear, Update, Delete & Table

+++ src/main/java/org/restaurant/category.java
@@ -5,10 +5,10 @@ public class category {
     private String productId;
     private String productName;
     private String type;
-    private int price;
+    private Double price;
     private String status;
 
-    public category(String productId, String productName, String type, int price, String status) {
+    public category(String productId, String productName, String type, Double price, String status) {
         this.productId = productId;
         this.productName = productName;
         this.type = type;
@@ -16,20 +16,15 @@ public class category {
         this.status = status;
     }
 
-    public String getProductId() {
-
-        return productId;
-    }
+    public String getProductId() {return productId;}
 
     public String getProductName() {
         return productName;
     }
 
-    public String getType() {
-        return type;
-    }
+    public String getType() {return type;}
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
